requirements.txt
openai
python-dotenv
pandas
pydantic
requests
beautifulsoup4
html2text
tiktoken
selenium
readability-lxml
streamlit
streamlit-tags
openpyxl
groq
google-generativeai
transformers
sentencepiece
sentence-transformers
jina
python=3.9